phpstan8

+++ src/models/synchronization/BaseEntity.php
@@ -266,7 +266,9 @@ class BaseEntity extends ActiveRecord
                         ]
                     )
                     ->one();
-                $column->delete();
+                if($column){
+                    $column->delete();
+                }
             }
         }
     }
